Refactored method checkUser! Now it return correct userId Refactored method delete user books when users was deleted move this method from booksRepo to usersRepo etc
