+++ .idea/modules.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project version="4">
+  <component name="ProjectModuleManager">
+    <modules>
+      <module fileurl="file://$PROJECT_DIR$/Server/Server.iml" filepath="$PROJECT_DIR$/Server/Server.iml" />
+      <module fileurl="file://$PROJECT_DIR$/ServerClient/ServerClient.iml" filepath="$PROJECT_DIR$/ServerClient/ServerClient.iml" />
+      <module fileurl="file://$PROJECT_DIR$/TestingMethods/TestingMethods.iml" filepath="$PROJECT_DIR$/TestingMethods/TestingMethods.iml" />
+    </modules>
+  </component>
+</project>
